fix memmove

+++ srcs/ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: kchen2 <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/13 14:34:59 by kchen2            #+#    #+#             */
-/*   Updated: 2019/03/06 14:12:01 by kchen2           ###   ########.fr       */
+/*   Updated: 2019/03/11 20:11:19 by kchen2           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ void	*ft_memmove(void *s1, const void *s2, size_t n)
 	c = (char *)s2;
 	b = (char *)s1;
 	if (c < b)
-		while ((--n) >= 0)
+		while ((int)(--n) >= 0)
 			*(b + n) = *(c + n);
 	else
 		while (i < n)
